Cross-platform CMake build (wxWidgets 3.2 + IM 3.15) (#1)
## Summary

Make the project build on Linux, macOS, and Windows from a single tree, with a unified Conan-based dependency story and a GitHub Actions matrix that runs on every PR. All three jobs are green.

- **Cross-platform CMake**: replace `linux/`, `osx/`, `win32/` with one root `CMakeLists.txt` driven by Conan-generated `find_package` targets.
- **Conan for everything**: a local recipe at `recipes/im/` builds the IM toolkit (which is not in any package manager) on Linux/macOS and downloads IM's official Win64 vc16 prebuilt zip on Windows; `conanfile.py` pulls wxWidgets 3.2, libxml2, libpng, libtiff, and FFTW from Conan Center. No more per-platform brew/apt/vcpkg gymnastics.
- **wxWidgets 2.8 → 3.2 source migration**: cmd-line struct types, `wxMenuItem` API renames, `wxCStrData` casts, `register` removal, an `in()` macro that collided with `std::codecvt::in()` once `<locale>` was pulled in, and `wxEVT_SCROLL_ENDSCROLL` → `wxEVT_SCROLL_CHANGED`.
- **IM 3.x → 3.15 source migration**: renamed headers, `imAnalyzeFindRegions` out-param, `imAnalyzeMeasureCentroid`/`Perimeter` `double*`, `imProcessToneGamut` params, `imCalcHistogram` taking `imImage*`.
- **Windows portability shims**:
  - `aruspix_uuid.h` exposes `ax_uuid_t` — a libuuid alias on POSIX, an `unsigned char[16]` array on Windows. Renamed because `<rpc.h>` already typedefs `uuid_t = UUID` and there is no way to keep both names in scope. `uuid_generate` is implemented out-of-line in `aruspix_uuid_win.cpp` against `UuidCreate` from `rpcrt4.lib`.
  - `<regex.h>` → C++ `<regex>` in `musiopae.cpp` (POSIX-extended flavour preserved).
  - Two `std::ifstream::open(wxString)` call sites disambiguated with `mb_str()` casts.
  - The legacy `__mb_cur_max`/`_pctype` MSVC link workaround in `imstaff.cpp` (now collides with the modern CRT) removed.
  - `WIN32_EXECUTABLE TRUE` so the linker uses `/SUBSYSTEM:WINDOWS` and the C runtime invokes wxWidgets' `WinMain` instead of looking for `int main()`.
  - `/FORCE:MULTIPLE /IGNORE:4006 /IGNORE:4088` because IM's prebuilt `im.lib` statically embeds libtiff/libpng/libjpeg/zlib, which collide with the same libraries wxWidgets pulls in transitively from Conan. The duplicates are functionally the same upstream code, so binding to either copy is safe.
  - FFTW forced to `shared=True` on Windows: IM's prebuilt `im_fftw3.lib` references `__imp_fftw_*` (DLL import-stub naming), so a static FFTW would not resolve.
- **CI**: single `.github/workflows/build.yml` with `matrix.os: [ubuntu-latest, macos-latest, windows-latest]`. Linux step lets Conan install GTK/X11/xcb dev packages itself (`tools.system.package_manager:mode=install`). Build parallelism is `2 × $(getconf _NPROCESSORS_ONLN)` (or `$NUMBER_OF_PROCESSORS` on Windows). Ninja generator forced everywhere so the same `conan-release` preset works on the three matrix entries.

## Bootstrap

```
pip install "conan>=2"
conan profile detect --force
conan export recipes/im
conan install . --build=missing
cmake --preset conan-release
cmake --build build/Release
```

On Linux add `-c tools.system.package_manager:mode=install -c tools.system.package_manager:sudo=True` to the `conan install` so Conan apt-installs the X11/GTK dev packages it needs (the workflow does this automatically).

## Notable IM patches in `recipes/im/conanfile.py`

- Linux/macOS only:
  - Delete `lz4/VERSION`, `libtiff/VERSION`, `libpng/VERSION` — case-insensitive APFS lets these shadow the C++ `<version>` header.
  - Route libpng include path through an env var → use Conan's libpng on every platform.
  - Same for FFTW (`im_fftw3` build).
  - Stop compiling IM's bundled libtiff `.c` sources into `libim.a` — they would collide with the libtiff that wxWidgets brings in transitively.
  - Build with `NO_DYNAMIC=Yes` so Tecmake skips its dylib link step.
- Windows: download the upstream `im-3.15_Win64_vc16_lib.zip` (vc16 is ABI-compatible with the vc17 MSVC toolchain on `windows-latest`); skip the source build entirely.

## Test plan

- [x] Local macOS Apple Silicon build (12 MB ARM64 binary)
- [x] CI Linux green
- [x] CI macOS green
- [x] CI Windows green (artifact uploaded)

+++ src/superimposition/supim.cpp
@@ -405,14 +405,14 @@ void SupImController::UpdateBrightness( )
     imImage *g_buf = imImageDuplicate( m_greenIm );
     imImage *imTmp = imImageCreate( w, h, IM_GRAY, IM_BYTE );
 
-    float param[2] = { 0, 0 }; // %
+    double param[2] = { 0, 0 }; // %
 
     if ( (m_greenBrightness != 0)  || (m_greenContrast != 0))
     {   
         param[0] = 5.0 * (float)m_greenBrightness;
         param[1] = 5.0 * (float)m_greenContrast;
         imProcessToneGamut( r_buf , imTmp, IM_GAMUT_BRIGHTCONT, param);
-        imProcessBitwiseOp( r_buf, g_buf, r_buf, IM_BIT_OR ); // valeurs communes doivent rester à 100%
+        imProcessBitwiseOp( r_buf, g_buf, r_buf, IM_BIT_OR ); // valeurs communes doivent rester ï¿½ 100%
         imProcessBitwiseOp( imTmp, r_buf, r_buf, IM_BIT_AND ); // AND entre valeurs communes et brightness ajuste
     }
     if ( (m_redBrightness != 0)  || (m_redContrast != 0))
@@ -420,7 +420,7 @@ void SupImController::UpdateBrightness( )
         param[0] = 5.0 * (float)m_redBrightness;
         param[1] = 5.0 * (float)m_redContrast;
         imProcessToneGamut( g_buf , imTmp, IM_GAMUT_BRIGHTCONT, param);
-        imProcessBitwiseOp( g_buf, r_buf, g_buf, IM_BIT_OR ); // valeurs communes doivent rester à 100%
+        imProcessBitwiseOp( g_buf, r_buf, g_buf, IM_BIT_OR ); // valeurs communes doivent rester ï¿½ 100%
         imProcessBitwiseOp( imTmp, g_buf, g_buf, IM_BIT_AND ); // AND entre valeurs communes et brightness ajuste
     }
     imProcessBitwiseOp( r_buf, g_buf, imTmp, IM_BIT_AND );
